Add CHW to local backend
CMSIS-NN and previous local backend is working on HWC for CPU acceleration.
However, AI chips such as K210's KPU is working in CHW format which is similar to most ML lib on the PC. A support for CHW format will let NNoM benefit for the hardware convolutions.

This commit add the CHW version to the below operators.
- Cropping
- Zero Padding
- Max, Avg, Sum pooling
- Up Sampling
Layers that need to add:
- Conv, Depthwise Conv
- Dense
Layers not affected (originally support both formats)
- Add, sub, mult
- All activations